requirements.txt
alembic==1.8.1
amqp==5.1.1
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
cryptography==38.0.1
Deprecated==1.2.13
Flask==2.2.2
Flask-Limiter==3.1.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==1.1.3.post0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
limits==3.1.5
Mako==1.2.3
MarkupSafe==2.1.1
ordered-set==4.1.0
packaging==23.0
prompt-toolkit==3.0.36
pycparser==2.21
Pygments==2.14.0
PyMySQL==1.0.2
python-dotenv==0.21.0
pytz==2022.7.1
redis==4.4.2
requests==2.28.1
rich==12.6.0
six==1.16.0
SQLAlchemy==1.4.42
typing_extensions==4.4.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
